refactor predict and parse
