+++ modules/modeling/module_modeling_llm/module_modeling_llm/core/filter_feedback.py
@@ -3,7 +3,7 @@ from langchain_core.prompts import ChatPromptTemplate
 
 from athena import emit_meta
 from module_modeling_llm.config import BasicApproachConfig
-from llm_core.utils.llm_utils import predict_and_parse
+from llm_core.utils.predict_and_parse import predict_and_parse
 from module_modeling_llm.models.assessment_model import AssessmentModel
 from module_modeling_llm.models.exercise_model import ExerciseModel
 from module_modeling_llm.prompts.filter_feedback_prompt import FilterFeedbackInputs
