+++ modules/programming/module_programming_llm/module_programming_llm/generate_non_graded_suggestions_by_file.py
@@ -15,8 +15,9 @@ from llm_core.utils.llm_utils import (
     check_prompt_length_and_omit_features_if_necessary,
     get_chat_prompt_with_formatting_instructions,
     num_tokens_from_string,
-    predict_and_parse,
 )
+from llm_core.utils.predict_and_parse import predict_and_parse
+
 from module_programming_llm.helpers.utils import (
     get_diff,
     load_files_from_repo,
