+++ modules/programming/module_programming_llm/module_programming_llm/generate_summary_by_file.py
@@ -12,8 +12,9 @@ from module_programming_llm.config import GradedBasicApproachConfig, BasicApproa
 from llm_core.utils.llm_utils import (
     get_chat_prompt_with_formatting_instructions,
     num_tokens_from_prompt,
-    predict_and_parse,
 )
+from llm_core.utils.predict_and_parse import predict_and_parse
+
 from module_programming_llm.helpers.utils import (
     get_diff,
     load_files_from_repo,
